fix the second copy of this file. XXX: merge them.

+++ sys/arch/mipsco/stand/common/saio.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: saio.c,v 1.13 2011/07/17 20:54:44 joerg Exp $	*/
+/*	$NetBSD: saio.c,v 1.14 2014/02/06 18:45:35 christos Exp $	*/
 
 /*
  * Copyright (c) 1992, 1993
@@ -150,7 +150,7 @@ saioopen(struct open_file *f, ...)
 	int i;
 	char *msg;
 	char buf[DEV_BSIZE];
-	int cnt;
+	size_t cnt;
 	static char device[] = "dksd(0,0,10)";
 
 	va_list ap;
